HuggingfaceAgent/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
huggingface_hub
transformers 
sentence-transformers
accelerate
langchain-text-splitters
langchain-chroma
chromadb
pypdf
